backend/requirements.txt
# Django and core dependencies
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.1.1

# Database
psycopg2-binary==2.9.11

# WebSocket support
channels==4.0.0

# Payment gateway
razorpay==1.4.2

# Development dependencies (optional)
# pytest==7.4.2
# pytest-django==4.8.0